@@ -9,6 +9,7 @@
     <modules>
         <module>kim-consumer</module>
         <module>kim-provider</module>
+        <module>kim-security</module>
     </modules>
     <name>kk-cloud</name>
     <description>Demo project for Spring Boot</description>
@@ -19,8 +20,8 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <maven.compiler.encoding>UTF-8</maven.compiler.encoding>
-        <spring-cloud.version>2021.0.2</spring-cloud.version>
-        <spring-boot.version>2.6.7</spring-boot.version>
+        <spring-cloud.version>2022.0.4</spring-cloud.version>
+        <spring-boot.version>3.1.3</spring-boot.version>
     </properties>
     <dependencies>
         <!--
@@ -65,22 +66,22 @@
         <dependency>
             <groupId>com.github.ben-manes.caffeine</groupId>
             <artifactId>caffeine</artifactId>
-            <version>3.1.1</version>
+            <version>3.1.5</version>
         </dependency>
 
-        <dependency>
+<!--        <dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-starter-sleuth</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-sleuth-zipkin</artifactId>
-        </dependency>
+        </dependency>-->
 
         <dependency>
             <groupId>net.logstash.logback</groupId>
             <artifactId>logstash-logback-encoder</artifactId>
-            <version>7.2</version>
+            <version>7.3</version>
         </dependency>
 
 
@@ -98,6 +99,7 @@
                 <groupId>org.springframework.cloud</groupId>
                 <artifactId>spring-cloud-dependencies</artifactId>
                 <version>${spring-cloud.version}</version>
+<!--                <version>2022.0.4</version>-->
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
