@@ -1,6 +1,7 @@
 package kim.kin.service;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
+
 import java.io.Serializable;
 
 /**
