@@ -1,5 +1,6 @@
 package kim.kin.rest;
 
+import jakarta.validation.Valid;
 import kim.kin.service.SendSmsReq;
 import kim.kin.service.SmsService;
 import kim.kin.service.TestService;
@@ -10,7 +11,6 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 
 @RestController
 public class TestController {
